refactored the code for sharedPreferences
